requirements.txt
selenium
webdriver-manager
openpyxl
pytest
pytest-html
allure-pytest